+++ packages/nimble-blazor/NimbleBlazor/WaferMapValidity.cs
@@ -7,7 +7,7 @@ public interface IWaferMapValidity
     public bool InvalidGridDimensions { get; }
 }
 
-internal class WaferMapValidity : IWaferMapValidity
+internal sealed class WaferMapValidity : IWaferMapValidity
 {
     [JsonPropertyName("invalidGridDimensions")]
     public bool InvalidGridDimensions { get; set; }