+++ packages/nimble-components/src/mapping/spinner/tests/mapping-spinner.stories.ts
@@ -26,8 +26,16 @@ export const spinnerMapping: StoryObj = {
             control: { type: 'none' }
         },
         text: {
+            control: { type: 'none' },
+            description: `A textual description of the value. The text will be displayed next to the spinner in a cell if \`text-hidden\` is not set,
+                or as the tooltip and accessible  name of the spinner if \`text-hidden\` is set. The text is also displayed next to the spinner
+                in a group header. This attribute is required.`
+        },
+        textHidden: {
+            control: { type: 'none' },
+            name: 'text-hidden',
             description:
-                'A textual description of the value which will be used as the tooltip and accessible name of the spinner. The text is also displayed next to the spinner in a group header. This attribute is required.'
+                "When set, the text is hidden within the table's cells. When unset, the text is rendered next to the spinner within the cell."
         }
     },
     args: {}