+++ packages/nimble-components/src/rich-text/models/rich-text-validator.ts
@@ -29,7 +29,7 @@ export class RichTextValidator {
         mentions: RichTextMention[]
     ): boolean {
         this.invalidMentionConfiguration = mentions.some(
-            x => !x.mentionInternals.validConfiguration
+            x => !x.mentionInternals.validator.isValid()
         );
         return !this.invalidMentionConfiguration;
     }