+++ packages/nimble-components/src/table-column/enum-base/index.ts
@@ -30,19 +30,14 @@ export interface TableColumnEnumColumnConfig
 }
 
 /**
- * Base class for table columns that map values to content (e.g. nimble-table-column-enum-text and nimble-table-column-icon)
+ * Base class for table columns that map values to content
  */
 export abstract class TableColumnEnumBase<
     TColumnConfig extends TableColumnEnumColumnConfig,
     TEnumValidator extends TableColumnEnumBaseValidator<[]>
 >
-    extends TableColumn<TColumnConfig>
+    extends TableColumn<TColumnConfig, TEnumValidator>
     implements Subscriber {
-    // To ensure the validator is available when other properties get initialized
-    // (which can trigger validation), declare the validator first.
-    /** @internal */
-    public validator = this.createValidator();
-
     /** @internal */
     public mappingNotifiers: Notifier[] = [];
 
@@ -68,8 +63,6 @@ export abstract class TableColumnEnumBase<
         }
     }
 
-    public abstract createValidator(): TEnumValidator;
-
     /**
      * Implementations should throw an error if an invalid Mapping is passed.
      */
@@ -85,8 +78,8 @@ export abstract class TableColumnEnumBase<
      * Called when any Mapping related state has changed.
      */
     private updateColumnConfig(): void {
-        this.validator.validate(this.mappings, this.keyType);
-        this.columnInternals.columnConfig = this.validator.isValid()
+        this.columnInternals.validator.validate(this.mappings, this.keyType);
+        this.columnInternals.columnConfig = this.checkValidity()
             ? this.createColumnConfig(this.getMappingConfigs())
             : undefined;
     }