+++ packages/nimble-components/src/table-column/enum-base/models/mapping-icon-config.ts
@@ -5,6 +5,7 @@ import { MappingConfig } from './mapping-config';
 export interface IconView {
     severity: IconSeverity;
     text?: string;
+    textHidden: boolean;
 }
 
 // Create an empty template containing only a space because creating a ViewTemplate
@@ -21,11 +22,11 @@ const createIconTemplate = (
 
     return html`
         <${icon}
-            title="${x => x.text}"
+            title="${x => (x.textHidden ? x.text : '')}"
             role="img"
             aria-label="${x => x.text}"
+            aria-hidden="${x => (x.textHidden ? 'false' : 'true')}"
             severity="${x => x.severity}"
-            class="no-shrink"
         >
         </${icon}>
     `;
@@ -36,10 +37,12 @@ const createIconTemplate = (
  */
 export class MappingIconConfig extends MappingConfig {
     public readonly iconTemplate: ViewTemplate<IconView>;
+
     public constructor(
         resolvedIcon: string | undefined,
         public readonly severity: IconSeverity,
-        text: string | undefined
+        text: string | undefined,
+        public readonly textHidden: boolean
     ) {
         super(text);
         this.iconTemplate = createIconTemplate(resolvedIcon);